+++ package.json
@@ -1,6 +1,7 @@
 {
+  "type": "module",
   "scripts": {
-    "start": "http-server -p 8080 && open http://localhost:8080"
+    "start": "http-server -p 8081 && open http://localhost:8081"
   },
   "devDependencies": {
     "http-server": "^14.1.1",
