+++ src/places.js
@@ -1,0 +1,7 @@
+const Place = {
+    HORSENS: "Horsens",
+    AARHUS: "Aarhus",
+    COPENHAGEN: "Copenhagen",
+  };
+
+  export default Place
