+++ src/types.js
@@ -1,0 +1,8 @@
+const Type = {
+    TEMPERATURE: "temperature",
+    PRECIPITATION: "precipitation",
+    WIND_SPEED: "wind speed",
+    CLOUD_COVERAGE: "cloud_coverage",
+  };
+
+  export default Type
