+++ src/units.js
@@ -1,0 +1,11 @@
+const Units = {
+    C: "C",
+    F: "F",
+    MM: "mm",
+    MS: "m/s",
+    PERCENT: "%",
+    INCHES: "inches",
+    MPH: "mph",
+  };
+
+  export default Units
